refactor(ui): extract heatmap + next-move-info builders
AIAnalysisContext.tsx: 592 → 355 lines. Two large pure useMemo bodies
moved to pure functions in contexts/ai/:

  - heatmapBuilder.ts: buildHeatmap(...) — assembles the on-board
    overlay from live MCTS topMoves or the final analysisResult, with
    the played-move ▷ marker overlay extracted into its own helper
  - nextMoveInfoBuilder.ts: computeNextMoveVertex() +
    computeNextMoveInfo() — look up the played move in AI suggestions,
    compute rank/winRate/scoreLead deltas

The provider now calls these from inside useMemo with the same deps,
so semantics and reactivity are unchanged.
